+++ web/sites/default/files/civicrm/ext/org.civicoop.emailapi/CRM/Emailapi/CivirulesAction/SendToContactReference.php
@@ -23,7 +23,7 @@ class CRM_Emailapi_CivirulesAction_SendToContactReference extends CRM_Civirules_
     $contactReferenceField = 'custom_' . $actionParams['contact_reference'];
     $entityData = $triggerData->getEntityData($actionParams['entity']);
     // Find the contact reference contact's ID.
-    $contactReferenceIds = (array) $entityData[$contactReferenceField];
+    $contactReferenceIds = (array) $entityData[$contactReferenceField.'_id'] ?? (array) $entityData[$contactReferenceField];
     // Not all $triggerData contains custom field data, so look it up if necessary.
     if (!$contactReferenceIds) {
       $contactReferenceIds = (array) civicrm_api3(CRM_Core_BAO_CustomGroup::getEntityFromExtends($actionParams['entity']), 'getvalue', [
@@ -72,26 +72,6 @@ class CRM_Emailapi_CivirulesAction_SendToContactReference extends CRM_Civirules_
   }
 
   /**
-   * Get a list of entities that use custom fields.
-   *
-   * @return array
-   */
-  public static function getContactReferenceEntities() {
-    $return[] = '-- please select --';
-    $result = \Civi\Api4\CustomField::get(TRUE)
-      ->addSelect('custom_group_id.extends')
-      ->addClause('OR', ['data_type', '=', 'ContactReference'], ['AND', [['data_type', '=', 'EntityReference'], ['fk_entity', '=', 'Contact']]])
-      ->execute()
-      ->indexBy('custom_group_id.extends');
-    foreach ($result as $field) {
-      $return[$field['custom_group_id.extends']] = $field['custom_group_id.extends'];
-    }
-    $return = array_unique($return);
-    asort($return);
-    return $return;
-  }
-
-  /**
    * Returns a redirect url to extra data input from the user after adding a action
    *
    * Return false if you do not need extra data input
@@ -111,15 +91,10 @@ class CRM_Emailapi_CivirulesAction_SendToContactReference extends CRM_Civirules_
    * @return string
    */
   public function userFriendlyConditionParams() {
-    $template = 'unknown template';
     $params = $this->getActionParameters();
 
-    $messageTemplates = new CRM_Core_DAO_MessageTemplate();
-    $messageTemplates->id = $params['template_id'];
-    $messageTemplates->is_active = TRUE;
-    if ($messageTemplates->find(TRUE)) {
-      $template = $messageTemplates->msg_title;
-    }
+    $template = CRM_Emailapi_CivirulesAction_Send::getTemplateLink($params['template_id']);
+
     if (isset($params['location_type_id']) && !empty($params['location_type_id'])) {
       try {
         $locationText = 'location type ' . civicrm_api3('LocationType', 'getvalue', [
@@ -153,7 +128,7 @@ class CRM_Emailapi_CivirulesAction_SendToContactReference extends CRM_Civirules_
     if (!empty($params['bcc'])) {
       $bcc = ts(' and bcc to %1', [1 => $params['bcc']]);
     }
-    return ts('Send e-mail from "%1 (%2 using %3)" with Template "%4" to %5 %6 %7', [
+    return ts("Send e-mail from '%1 (%2 using %3)' with Template '%4' to %5 %6 %7", [
       1 => $params['from_name'],
       2 => $params['from_email'],
       3 => $locationText,