+++ web/sites/default/files/civicrm/ext/org.civicoop.emailapi/CRM/Emailapi/Form/CivirulesAction/SendToContactReference.php
@@ -7,8 +7,35 @@
  */
 class CRM_Emailapi_Form_CivirulesAction_SendToContactReference extends CRM_Emailapi_Form_CivirulesAction_Send {
 
-  protected function getContactReferenceEntities() {
-    return CRM_Emailapi_CivirulesAction_SendToContactReference::getContactReferenceEntities();
+  /**
+   * Get a list of entities that use custom fields.
+   *
+   * @return array
+   */
+  public function getContactReferenceEntities() {
+
+    $entities = [];
+    foreach ($this->triggerClass->getProvidedEntities() as $entityDef) {
+      $entity = $entityDef->entity;
+      if ($entity == 'Contact') {
+        $entities += ['Contact', 'Individual', 'Organization', 'Household'];
+      } else {
+        $entities[] = $entity;
+      }
+    }
+    $return[] = '-- please select --';
+    $result = \Civi\Api4\CustomField::get(TRUE)
+      ->addSelect('custom_group_id.extends')
+      ->addWhere('custom_group_id.extends', 'IN', $entities)
+      ->addClause('OR', ['data_type', '=', 'ContactReference'], ['AND', [['data_type', '=', 'EntityReference'], ['fk_entity', '=', 'Contact']]])
+      ->execute()
+      ->indexBy('custom_group_id.extends');
+    foreach ($result as $field) {
+      $return[$field['custom_group_id.extends']] = $field['custom_group_id.extends'];
+    }
+    $return = array_unique($return);
+    asort($return);
+    return $return;
   }
 
   public function buildQuickForm() {
@@ -57,5 +84,7 @@ class CRM_Emailapi_Form_CivirulesAction_SendToContactReference extends CRM_Email
     $data['contact_reference'] = $this->_submitValues['contact_reference'];
     parent::postProcess($data);
   }
+
+
 
 }